create Upload pic hook
