+++ src/pages/Profile/components/MyQR/EditQRComponent.tsx
@@ -5,6 +5,8 @@ import { userInfoStore } from '../../../../store/user.store'
 import { firebaseApp } from '../../../../config/firebase'
 import { dontSubmitWhenEnter } from '../../../../utils/eventManage'
 import { uploadImage } from '../../../../service/storage'
+import { UploadFile } from 'antd/lib/upload/interface'
+import { UploadPic } from '../hook'
 
 export interface UpdateValue {
   aboutme: string
@@ -15,48 +17,21 @@ export interface EditComponentProps {
   onClose: () => void
 }
 
-/* TODO: upload QR picture function*/
-
 export const EditQRComponent: React.FC<EditComponentProps> = props => {
   const [isUploading, setIsUploading] = useState(false)
   const { userInfo } = userInfoStore()
   const { TextArea } = Input
   const { onClose } = props
-  const [QRUrl, setQRUrl] = useState(String)
-
-  const uploadNewImage = async (file: File) => {
-    try {
-      setIsUploading(true)
-      const uploadStatus = await uploadImage(file)
-      if (uploadStatus.url) {
-        setQRUrl(uploadStatus.url)
-      }
-    } catch (error: any) {
-      console.log(error)
-    }
-  }
-
-  const handleRequest = (option: any) => {
-    uploadNewImage(option.file).finally(() => setIsUploading(false))
-  }
+  const [imageUrl, setimageUrl] = useState(String)
+  const { handleRequest, beforeUpload } = UploadPic({ setimageUrl, setIsUploading })
 
   const onFinish = (value: UpdateValue) => {
     onClose()
     console.log(value)
+    console.log(imageUrl)
     /* TODO: update QR */
   }
 
-  function beforeUpload(file: any) {
-    const isJpgOrPng = file.type === 'image/jpeg' || file.type === 'image/png'
-    if (!isJpgOrPng) {
-      message.error('You can only upload JPG/PNG file!')
-    }
-    const isLt2M = file.size / 1024 / 1024 < 2
-    if (!isLt2M) {
-      message.error('Image must smaller than 2MB!')
-    }
-    return isJpgOrPng && isLt2M
-  }
   return (
     <div className="p-3">
       <Divider>
@@ -82,7 +57,7 @@ export const EditQRComponent: React.FC<EditComponentProps> = props => {
             beforeUpload={beforeUpload}
             showUploadList={false}
           >
-            {QRUrl ? <img src={QRUrl} alt="QR" /> : <p>Upload</p>}
+            {imageUrl ? <img src={imageUrl} alt="QR" /> : <p>Upload</p>}
           </Upload>
         </Form.Item>
         <Form.Item name="aboutdonate">
